Save as /etc/ssh/ssh_known_hosts (or /root/.ssh/known_hosts) This file exists to make sure that you are really connecting to the official studnet server, and not a faked one!
For the case that this guide will be outdated one day, you can create it this way:
Connect manually to the studnet with "ssh mieternummer@serverip".
You will be asked if the fingerprint is correct. Compare it with the one the studnet released.
Really, check that key once. Because it could also mean that someone fakes the studnet server.
Feel free to create a Pull Request or Issue on github afterwards ;)